docs(example): more refactoring in prep to move to sdk
