+++ examples/authentication/regular-web-app/lib/middleware/hasteUserInViews.js
