+++ examples/authentication/regular-web-app/routes/user.js
@@ -6,9 +6,7 @@ const router = Router();
 
 /* GET user profile. */
 router.get('/user', secured(), function (req, res, next) {
-  const { _raw, _json, ...userProfile } = req.user;
   res.render('index', {
-    userProfile: JSON.stringify(userProfile, null, 2),
     title: 'Profile page',
   });
 });
